Add badge for cart
now the badge will calculate at the first load of the app
on every app count change. a dedicated function will be called to change
the cart count in badge.

for increment and decrement, since we don't disable add to cart button
in order for states to update correctly, i used an updater function
instead of normal state setting.